pyproject.toml
[project]
name = "whisper-openapi"
description = "Whisper ASR OpenAI API Server."
version = "0.1.0"
readme = "README.md"
requires-python = "<3.13,>=3.10"
dependencies = [
"fastapi >=0.115.8",
"uvicorn >=0.34.0" ,
"python-multipart >=0.0.20",
"ffmpeg-python >=0.2.0",
"numpy <= 2.0.0",
"openai-whisper >=20240930",
"faster-whisper >=1.1.0",
"whisperx >=3.3.1",
"tqdm >=4.67.1",
"llvmlite >=0.44.0",
"numba >=0.61.0"
]


[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.uv]
cache-dir = "/data/.uv_cache"

